+++ packages/components/src/components/CheckboxGroup/stories/Default.stories.tsx
@@ -69,3 +69,17 @@ export const WithFieldError: Story = {
     </CheckboxGroup>
   ),
 };
+
+export const ColumnLayout: Story = {
+  render: (props) => (
+    <CheckboxGroup l={[1, 1, 1]} m={[1, 1]} {...props}>
+      <Label>Options</Label>
+      <Checkbox value="1">Option 1</Checkbox>
+      <Checkbox value="2">Option 2</Checkbox>
+      <Checkbox value="3">Option 3</Checkbox>
+      <Checkbox value="4">Option 4</Checkbox>
+      <Checkbox value="5">Option 5</Checkbox>
+      <Checkbox value="6">Option 6</Checkbox>
+    </CheckboxGroup>
+  ),
+};