+++ packages/remote-elements/src/lib/getObjectKeysIncludingProtoTypes.ts
@@ -1,9 +1,0 @@
-export const getObjectKeysIncludingProtoTypes = (something: object) => {
-  const result: string[] = [];
-
-  for (const key in something) {
-    result.push(key);
-  }
-
-  return result;
-};