+++ CHANGELOG.md
@@ -3,6 +3,53 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.237](https://github.com/mittwald/flow/compare/0.2.0-alpha.236...0.2.0-alpha.237) (2025-06-04)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.236](https://github.com/mittwald/flow/compare/0.2.0-alpha.235...0.2.0-alpha.236) (2025-06-04)
+
+### Bug Fixes
+
+* Allow control prop ([#1557](https://github.com/mittwald/flow/issues/1557)) ([c5ce6f0](https://github.com/mittwald/flow/commit/c5ce6f03d28500ae335067c70ac84e2632517567))
+* **Option:** use value for key to support remote rendering ([#1556](https://github.com/mittwald/flow/issues/1556)) ([3fea5c8](https://github.com/mittwald/flow/commit/3fea5c8b9a68f7d71740b8dfff399cb87413a048))
+
+# [0.2.0-alpha.235](https://github.com/mittwald/flow/compare/0.2.0-alpha.234...0.2.0-alpha.235) (2025-06-03)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.234](https://github.com/mittwald/flow/compare/0.2.0-alpha.233...0.2.0-alpha.234) (2025-06-03)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.233](https://github.com/mittwald/flow/compare/0.2.0-alpha.232...0.2.0-alpha.233) (2025-06-03)
+
+### Bug Fixes
+
+* **Message:** update border-radius ([#1561](https://github.com/mittwald/flow/issues/1561)) ([07a9745](https://github.com/mittwald/flow/commit/07a974510be9eef8e96b36aceb62d68a264ec045))
+
+# [0.2.0-alpha.232](https://github.com/mittwald/flow/compare/0.2.0-alpha.231...0.2.0-alpha.232) (2025-06-03)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.231](https://github.com/mittwald/flow/compare/0.2.0-alpha.230...0.2.0-alpha.231) (2025-06-03)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.230](https://github.com/mittwald/flow/compare/0.2.0-alpha.229...0.2.0-alpha.230) (2025-06-03)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.229](https://github.com/mittwald/flow/compare/0.2.0-alpha.228...0.2.0-alpha.229) (2025-06-03)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.228](https://github.com/mittwald/flow/compare/0.2.0-alpha.227...0.2.0-alpha.228) (2025-06-02)
+
+### Bug Fixes
+
+* **CopyButton:** fix not copied to clipboard in remote context ([#1550](https://github.com/mittwald/flow/issues/1550)) ([7aa03a5](https://github.com/mittwald/flow/commit/7aa03a5d7559bce6fcb834cd11411f4de6d6f8d0))
+
 # [0.2.0-alpha.227](https://github.com/mittwald/flow/compare/0.2.0-alpha.226...0.2.0-alpha.227) (2025-05-30)
 
 ### Features